[#38] refactor: Change recordList location to organisms
